Improve performance.
